rename Contract container to EmpContract
