+++ pages/_app.tsx
@@ -5,7 +5,7 @@ import "../utils/global.css";
 
 import { WithStylingProviders } from "../utils/styling";
 import Connection from "../containers/Connection";
-import Contract from "../containers/Contract";
+import EmpContract from "../containers/EmpContract";
 import EmpAddress from "../containers/EmpAddress";
 import EmpState from "../containers/EmpState";
 import Collateral from "../containers/Collateral";
@@ -20,7 +20,7 @@ interface IProps {
 const WithStateContainerProviders = ({ children }: IProps) => (
   <Connection.Provider>
     <EmpAddress.Provider>
-      <Contract.Provider>
+      <EmpContract.Provider>
         <EmpState.Provider>
           <Collateral.Provider>
             <Token.Provider>
@@ -30,7 +30,7 @@ const WithStateContainerProviders = ({ children }: IProps) => (
             </Token.Provider>
           </Collateral.Provider>
         </EmpState.Provider>
-      </Contract.Provider>
+      </EmpContract.Provider>
     </EmpAddress.Provider>
   </Connection.Provider>
 );
